refactor: move logger to daggers package
