--- dagger/logger/doc.go
@@ -1,4 +1,0 @@
-// Package logger provides a simple logging interface for use in the dagger client.
-//
-// The logger is a simple interface wrapper interface around the standard io.Writer interface.
-package logger
