+++ daggers/logger.go
@@ -1,4 +1,4 @@
-package logger
+package daggers
 
 import (
 	"fmt"
